+++ jcore/src/main/java/io/memoria/jutils/jcore/msgbus/MsgBusAdmin.java
@@ -5,7 +5,7 @@ import reactor.core.publisher.Mono;
 public interface MsgBusAdmin {
   Mono<Void> createTopic(String topic, int partitions, int replicationFactor);
 
-  Mono<Integer> currentOffset(String topic, int partition);
+  Mono<Long> currentOffset(String topic, int partition);
 
   Mono<Boolean> exists(String topic);
